+++ package.json
@@ -11,13 +11,8 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
-		"@hapi/joi": "^17.1.1",
-		"bcryptjs": "^2.4.3",
 		"body-parser": "^1.19.0",
-		"dotenv": "^8.2.0",
 		"express": "^4.17.1",
-		"joi": "^17.4.0",
-		"jsonwebtoken": "^8.5.1",
 		"mongoose": "^5.11.19",
 		"nodemon": "^2.0.7"
 	}
